bot_backup/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-chroma
langchain-huggingface
langchain
pydantic
